update admin controllers and blades
