+++ app/Http/Controllers/Admin/IndexController.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Http\Controllers\Admin;
+
+use Illuminate\Http\Request;
+
+use App\Http\Requests;
+use App\Http\Controllers\Controller;
+
+class IndexController extends Controller
+{
+    public function index()
+    {
+        return view('admin/Index');
+    }
+}
